refactor: ScheduleEventMapper 제거 및 확장 함수 위치 변경
